Fix jobcontrol bug

+++ srcs/jobcontrol/job_catch_sigchld.c
@@ -6,13 +6,14 @@
 /*   By: frossiny <frossiny@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/29 14:36:45 by frossiny          #+#    #+#             */
-/*   Updated: 2020/01/20 11:13:47 by frossiny         ###   ########.fr       */
+/*   Updated: 2020/01/22 17:15:34 by frossiny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "shell.h"
 #include "jobcontrol.h"
 #include "execution.h"
+#include "ast.h"
 
 static void	catch_gchild(void)
 {
@@ -20,7 +21,7 @@ static void	catch_gchild(void)
 	int			pid;
 
 	pid = waitpid(g_child, &g_lstatus, WNOHANG);
-	if (pid == 0)
+	if (pid == 0 && g_shell.current_cmd->exe_found)
 	{
 		kill(-g_child, SIGTSTP);
 		if (g_shell.current_cmd)
@@ -32,7 +33,7 @@ static void	catch_gchild(void)
 		}
 		job->status = "Stopped";
 		job->state = JOB_SUSPENDED;
-		g_shell.current_cmd = NULL;
+		ast_free_cmd(g_shell.current_cmd);
 		g_child = 0;
 		tcsetpgrp(g_shell.pgrp, g_shell.pid);
 	}
@@ -46,9 +47,9 @@ void		job_catch_sigchld(int signal)
 	int			status;
 
 	(void)signal;
-	if (g_child)
+	if (g_child && g_shell.current_cmd)
 		catch_gchild();
-	else if (!g_shell.current_pipel)
+	else if (!g_shell.current_pipel && !g_shell.current_cmd)
 	{
 		pid = waitpid(0, &status, WNOHANG);
 		if (pid > 0)
